make render branch stack generation handle 2+ inputs
While a maximum of 2 image inputs is supported, branchForNode can construct a trees with >2 inputs and the inner render loop within evaluate can assign input values for >2 inputs. To follow suit, renderStackForRootBranch should now be able to handle >2 inputs as well.

That said, the FBO is currently fixed at 3 attachments which will only support a maximum of 2 inputs following what I think the calculation to be:
  attachments = max inputs + 1
which allows 2 attachments for input images and another to write the result to.