improve and simply dependency solver
The render stack generation with the previous dependency solver failed in many pipeline arrangements. The new solver has a 3 pass stratagem that:
[1] Walks up the graph from the render node taking note of which node connections are required and also the required source (leaf) nodes.
[2] Walks down the graph from the source nodes to the render node, recording the maximum output cost of each of the required connections.
[3] Again walks up the graph from the render node, pushing nodes onto a stack, preferring the cheaper connections first.

Some changes were made in the context evaluation as well, instead of trickling an enum through the evaluation process of out to in attachments, the attachment that each rendered node is drawn to is tracked to allow more flexibility. This likely introduces some additional overhead since a map tracks all render node > attachment, but it greatly simplifies the code which is a big plus.